fix feedback issues + api get titles

+++ database/migrations/2023_01_27_155826_create_feedbacks_courses_table.php
@@ -13,30 +13,32 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('feedbacks_courses', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('feedbacks_courses')) {
+            Schema::create('feedbacks_courses', function (Blueprint $table) {
+                $table->id();
 
-            $table->bigInteger('id_course')->unsigned();
-            $table->foreign('id_course')
-                ->references('id')
-                ->on('courses')
-                ->constrained()
-                ->onDelete('cascade');
+                $table->bigInteger('id_course')->unsigned();
+                $table->foreign('id_course')
+                    ->references('id')
+                    ->on('courses')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->bigInteger('id_feedback')->unsigned();
-            $table->foreign('id_feedback')
-                ->references('id')
-                ->on('feedbacks')
-                ->constrained()
-                ->onDelete('cascade');
+                $table->bigInteger('id_feedback')->unsigned();
+                $table->foreign('id_feedback')
+                    ->references('id')
+                    ->on('feedbacks')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->bigInteger('id_user')->unsigned();
-            $table->foreign('id_user')
-                ->references('id')
-                ->on('users')
-                ->constrained()
-                ->onDelete('cascade');
-        });
+                $table->bigInteger('id_user')->unsigned();
+                $table->foreign('id_user')
+                    ->references('id')
+                    ->on('users')
+                    ->constrained()
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
